nlp-server/requirements.txt
fastapi
uvicorn
openai
langchain
python-dotenv
langchain-openai
email-validator
